Enable even more fine-grained plugin reloading
