+++ packages/docusaurus/src/server/plugins/plugins.ts
@@ -5,14 +5,19 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import _ from 'lodash';
-import logger from '@docusaurus/logger';
 import {initPlugins} from './init';
 import {createBootstrapPlugin, createMDXFallbackPlugin} from './synthetic';
 import {localizePluginTranslationFile} from '../translations/translations';
 import {sortRoutes} from './routeConfig';
 import {PerfLogger} from '../../utils';
 import {createPluginActionsUtils} from './actions';
+import {
+  aggregateAllContent,
+  aggregateGlobalData,
+  aggregateRoutes,
+  getPluginByIdentifier,
+  mergeGlobalData,
+} from './pluginsUtils';
 import type {
   LoadContext,
   RouteConfig,
@@ -23,17 +28,17 @@ import type {
   InitializedPlugin,
 } from '@docusaurus/types';
 
-async function translatePlugin({
+async function translatePluginContent({
   plugin,
+  content,
   context,
 }: {
-  plugin: LoadedPlugin;
+  plugin: InitializedPlugin;
+  content: unknown;
   context: LoadContext;
-}): Promise<LoadedPlugin> {
-  const {content} = plugin;
-
+}): Promise<unknown> {
   const rawTranslationFiles =
-    (await plugin.getTranslationFiles?.({content: plugin.content})) ?? [];
+    (await plugin.getTranslationFiles?.({content})) ?? [];
 
   const translationFiles = await Promise.all(
     rawTranslationFiles.map((translationFile) =>
@@ -58,10 +63,10 @@ async function translatePlugin({
   // translate its own slice of theme config and should make no assumptions
   // about other plugins' keys, so this is safe to run in parallel.
   Object.assign(context.siteConfig.themeConfig, translatedThemeConfigSlice);
-  return {...plugin, content: translatedContent};
+  return translatedContent;
 }
 
-async function executePluginLoadContent({
+async function executePluginContentLoading({
   plugin,
   context,
 }: {
@@ -69,70 +74,59 @@ async function executePluginLoadContent({
   context: LoadContext;
 }): Promise<LoadedPlugin> {
   return PerfLogger.async(
-    `Plugin - loadContent - ${plugin.name}@${plugin.options.id}`,
+    `Plugins - single plugin content loading - ${plugin.name}@${plugin.options.id}`,
     async () => {
-      const content = await plugin.loadContent?.();
-      const loadedPlugin: LoadedPlugin = {...plugin, content};
-      return translatePlugin({plugin: loadedPlugin, context});
-    },
-  );
-}
+      let content = await plugin.loadContent?.();
 
-async function executePluginsLoadContent({
-  plugins,
-  context,
-}: {
-  plugins: InitializedPlugin[];
-  context: LoadContext;
-}) {
-  return PerfLogger.async(`Plugins - loadContent`, () =>
-    Promise.all(
-      plugins.map((plugin) => executePluginLoadContent({plugin, context})),
-    ),
-  );
-}
+      content = await translatePluginContent({
+        plugin,
+        content,
+        context,
+      });
 
-function aggregateAllContent(loadedPlugins: LoadedPlugin[]): AllContent {
-  return _.chain(loadedPlugins)
-    .groupBy((item) => item.name)
-    .mapValues((nameItems) =>
-      _.chain(nameItems)
-        .groupBy((item) => item.options.id)
-        .mapValues((idItems) => idItems[0]!.content)
-        .value(),
-    )
-    .value();
-}
-
-async function executePluginContentLoaded({
-  plugin,
-  context,
-}: {
-  plugin: LoadedPlugin;
-  context: LoadContext;
-}): Promise<{routes: RouteConfig[]; globalData: unknown}> {
-  return PerfLogger.async(
-    `Plugins - contentLoaded - ${plugin.name}@${plugin.options.id}`,
-    async () => {
       if (!plugin.contentLoaded) {
-        return {routes: [], globalData: undefined};
+        return {
+          ...plugin,
+          content,
+          routes: [],
+          globalData: undefined,
+        };
       }
+
       const pluginActionsUtils = await createPluginActionsUtils({
         plugin,
         generatedFilesDir: context.generatedFilesDir,
         baseUrl: context.siteConfig.baseUrl,
         trailingSlash: context.siteConfig.trailingSlash,
       });
+
       await plugin.contentLoaded({
-        content: plugin.content,
+        content,
         actions: pluginActionsUtils.getActions(),
       });
+
       return {
+        ...plugin,
+        content,
         routes: pluginActionsUtils.getRoutes(),
         globalData: pluginActionsUtils.getGlobalData(),
       };
     },
   );
+}
+
+async function executeAllPluginsContentLoading({
+  plugins,
+  context,
+}: {
+  plugins: InitializedPlugin[];
+  context: LoadContext;
+}): Promise<LoadedPlugin[]> {
+  return PerfLogger.async(`Plugins - all plugins content loading`, () => {
+    return Promise.all(
+      plugins.map((plugin) => executePluginContentLoading({plugin, context})),
+    );
+  });
 }
 
 async function executePluginAllContentLoaded({
@@ -168,49 +162,15 @@ async function executePluginAllContentLoaded({
   );
 }
 
-async function executePluginsContentLoaded({
+type AllContentLoadedResult = {routes: RouteConfig[]; globalData: GlobalData};
+
+async function executeAllPluginsAllContentLoaded({
   plugins,
   context,
 }: {
   plugins: LoadedPlugin[];
   context: LoadContext;
-}): Promise<{routes: RouteConfig[]; globalData: GlobalData}> {
-  return PerfLogger.async(`Plugins - contentLoaded`, async () => {
-    const routes: RouteConfig[] = [];
-    const globalData: GlobalData = {};
-
-    await Promise.all(
-      plugins.map(async (plugin) => {
-        const {routes: pluginRoutes, globalData: pluginGlobalData} =
-          await executePluginContentLoaded({
-            plugin,
-            context,
-          });
-
-        routes.push(...pluginRoutes);
-
-        if (pluginGlobalData !== undefined) {
-          globalData[plugin.name] ??= {};
-          globalData[plugin.name]![plugin.options.id] = pluginGlobalData;
-        }
-      }),
-    );
-
-    // Sort the route config.
-    // This ensures that route with sub routes are always placed last.
-    sortRoutes(routes, context.siteConfig.baseUrl);
-
-    return {routes, globalData};
-  });
-}
-
-async function executePluginsAllContentLoaded({
-  plugins,
-  context,
-}: {
-  plugins: LoadedPlugin[];
-  context: LoadContext;
-}): Promise<{routes: RouteConfig[]; globalData: GlobalData}> {
+}): Promise<AllContentLoadedResult> {
   return PerfLogger.async(`Plugins - allContentLoaded`, async () => {
     const allContent = aggregateAllContent(plugins);
 
@@ -235,12 +195,29 @@ async function executePluginsAllContentLoaded({
       }),
     );
 
-    // Sort the route config.
-    // This ensures that route with sub routes are always placed last.
-    sortRoutes(routes, context.siteConfig.baseUrl);
-
     return {routes, globalData};
   });
+}
+
+function mergeResults({
+  plugins,
+  allContentLoadedResult,
+}: {
+  plugins: LoadedPlugin[];
+  allContentLoadedResult: AllContentLoadedResult;
+}) {
+  const routes: RouteConfig[] = [
+    ...aggregateRoutes(plugins),
+    ...allContentLoadedResult.routes,
+  ];
+  sortRoutes(routes);
+
+  const globalData: GlobalData = mergeGlobalData(
+    aggregateGlobalData(plugins),
+    allContentLoadedResult.globalData,
+  );
+
+  return {routes, globalData};
 }
 
 export type LoadPluginsResult = {
@@ -248,52 +225,6 @@ export type LoadPluginsResult = {
   routes: RouteConfig[];
   globalData: GlobalData;
 };
-
-type ContentLoadedResult = {routes: RouteConfig[]; globalData: GlobalData};
-
-export function mergeGlobalData(...globalDataList: GlobalData[]): GlobalData {
-  const result: GlobalData = {};
-
-  const allPluginIdentifiers: PluginIdentifier[] = globalDataList.flatMap(
-    (gd) =>
-      Object.keys(gd).flatMap((name) =>
-        Object.keys(gd[name]!).map((id) => ({name, id})),
-      ),
-  );
-
-  allPluginIdentifiers.forEach(({name, id}) => {
-    const allData = globalDataList
-      .map((gd) => gd?.[name]?.[id])
-      .filter((d) => typeof d !== 'undefined');
-    const mergedData =
-      allData.length === 1 ? allData[0] : Object.assign({}, ...allData);
-    result[name] ??= {};
-    result[name]![id] = mergedData;
-  });
-
-  return result;
-}
-
-function mergeResults({
-  contentLoadedResult,
-  allContentLoadedResult,
-}: {
-  contentLoadedResult: ContentLoadedResult;
-  allContentLoadedResult: ContentLoadedResult;
-}): ContentLoadedResult {
-  const routes = [
-    ...contentLoadedResult.routes,
-    ...allContentLoadedResult.routes,
-  ];
-  sortRoutes(routes);
-
-  const globalData = mergeGlobalData(
-    contentLoadedResult.globalData,
-    allContentLoadedResult.globalData,
-  );
-
-  return {routes, globalData};
-}
 
 /**
  * Initializes the plugins and run their lifecycle functions.
@@ -307,52 +238,29 @@ export async function loadPlugins(
       () => initPlugins(context),
     );
 
+    // TODO probably not the ideal place to hardcode those plugins
     initializedPlugins.push(
       createBootstrapPlugin(context),
       createMDXFallbackPlugin(context),
     );
 
-    const plugins = await executePluginsLoadContent({
+    const plugins = await executeAllPluginsContentLoading({
       plugins: initializedPlugins,
       context,
     });
 
-    const contentLoadedResult = await executePluginsContentLoaded({
-      plugins,
-      context,
-    });
-
-    const allContentLoadedResult = await executePluginsAllContentLoaded({
+    const allContentLoadedResult = await executeAllPluginsAllContentLoaded({
       plugins,
       context,
     });
 
     const {routes, globalData} = mergeResults({
-      contentLoadedResult,
+      plugins,
       allContentLoadedResult,
     });
 
     return {plugins, routes, globalData};
   });
-}
-
-export function getPluginByIdentifier({
-  plugins,
-  pluginIdentifier,
-}: {
-  pluginIdentifier: PluginIdentifier;
-  plugins: LoadedPlugin[];
-}): LoadedPlugin {
-  const plugin = plugins.find(
-    (p) =>
-      p.name === pluginIdentifier.name && p.options.id === pluginIdentifier.id,
-  );
-  if (!plugin) {
-    throw new Error(
-      logger.interpolate`Plugin not found for identifier ${pluginIdentifier.name}@${pluginIdentifier.id}`,
-    );
-  }
-  return plugin;
 }
 
 export async function reloadPlugin({
@@ -365,30 +273,26 @@ export async function reloadPlugin({
   context: LoadContext;
 }): Promise<LoadPluginsResult> {
   return PerfLogger.async('Plugins - reloadPlugin', async () => {
-    const plugin = getPluginByIdentifier({
+    const previousPlugin = getPluginByIdentifier({
       plugins: previousPlugins,
       pluginIdentifier,
     });
-
-    const reloadedPlugin = await executePluginLoadContent({plugin, context});
-    const plugins = previousPlugins.with(
-      previousPlugins.indexOf(plugin),
-      reloadedPlugin,
-    );
-
-    // TODO optimize this, we shouldn't need to re-run this lifecycle
-    const contentLoadedResult = await executePluginsContentLoaded({
-      plugins,
+    const plugin = await executePluginContentLoading({
+      plugin: previousPlugin,
       context,
     });
+    const plugins = previousPlugins.with(
+      previousPlugins.indexOf(previousPlugin),
+      plugin,
+    );
 
-    const allContentLoadedResult = await executePluginsAllContentLoaded({
+    const allContentLoadedResult = await executeAllPluginsAllContentLoaded({
       plugins,
       context,
     });
 
     const {routes, globalData} = mergeResults({
-      contentLoadedResult,
+      plugins,
       allContentLoadedResult,
     });
